Improve output, logging, and cleanup.
